Move API submodules to ApiSubmodules subdirectory; fix module exports
