+++ src/index.ts
@@ -1,4 +1,31 @@
-import PointApi, { Account, ErrorResponse, Subscription } from "./main";
-import { Suggestion, Hotkey } from "./customSuggestions";
+import PointApi, { ErrorResponse, Account, Subscription } from "./main";
+import AutocompleteSession, {
+  ContextType,
+  SuggestionMeta,
+  ReplyMeta,
+  AutocompleteResponse,
+  ReplyResponse
+} from "./ApiModules/autocompleteSession";
+
+import CustomSuggestionsApiModule, {
+  Suggestion,
+  Hotkey
+} from "./ApiModules/customSuggestions";
+
+import InteractionsApiModule, {
+  StatusResponse
+} from "./ApiModules/interactions";
+
 export default PointApi;
-export { Account, ErrorResponse, Subscription, Suggestion, Hotkey };
+export { ErrorResponse, Account, Subscription };
+export {
+  AutocompleteSession,
+  ContextType,
+  SuggestionMeta,
+  ReplyMeta,
+  AutocompleteResponse,
+  ReplyResponse
+};
+export { CustomSuggestionsApiModule, Suggestion, Hotkey };
+export { InteractionsApiModule, StatusResponse };
+export * from "./main";
